Feat: FileQueryService - getStickerImgMap
